refactor: organizing logging and middleware into subpackages
