+++ services/tracker/src/tracker/logging/__init__.py
@@ -1,0 +1,20 @@
+"""Structured logging subpackage for the tracker service."""
+
+from tracker.logging.config import DevFormatter, configure_logging
+from tracker.logging.context import (
+    ContextFilter,
+    benchmark_id_var,
+    request_id_var,
+    task_id_var,
+)
+from tracker.logging.logger import get_logger
+
+__all__ = [
+    "ContextFilter",
+    "DevFormatter",
+    "benchmark_id_var",
+    "configure_logging",
+    "get_logger",
+    "request_id_var",
+    "task_id_var",
+]
