+++ services/tracker/src/tracker/middleware/__init__.py
@@ -1,0 +1,11 @@
+"""Middleware subpackage for the tracker service."""
+
+from tracker.middleware.logging_context import LoggingContextMiddleware
+from tracker.middleware.request_context import RequestContextMiddleware
+from tracker.middleware.task_protection import TaskProtectionMiddleware
+
+__all__ = [
+    "LoggingContextMiddleware",
+    "RequestContextMiddleware",
+    "TaskProtectionMiddleware",
+]
