+++ services/tracker/src/tracker/sandbox.py
@@ -38,7 +38,7 @@ from tenacity import (
 
 from tracker.database.models import AgentContractRequest
 from tracker.exceptions import InvalidSandboxConfigurationError, SandboxError
-from tracker.logger import get_logger
+from tracker.logging import get_logger
 from tracker.s3 import download_from_s3, get_contract_s3_key, upload_to_s3
 from tracker.types import AWSCredentials
